+++ src/promptflow-core/tests/conftest.py
@@ -1,5 +1,7 @@
 from pathlib import Path
 
+from promptflow._utils.flow_utils import resolve_flow_path
+
 TEST_CONFIG_ROOT = Path(__file__).parent.parent.parent / "promptflow" / "tests" / "test_configs"
 FLOW_ROOT = TEST_CONFIG_ROOT / "flows"
 EAGER_FLOW_ROOT = TEST_CONFIG_ROOT / "eager_flows"
@@ -10,6 +12,11 @@ def get_flow_folder(folder_name, root: str = FLOW_ROOT) -> Path:
     return flow_folder_path
 
 
-def get_yaml_file(folder_name, root: str = FLOW_ROOT, file_name: str = "flow.dag.yaml") -> Path:
-    yaml_file = get_flow_folder(folder_name, root) / file_name
+def get_yaml_file(folder_name, root: str = FLOW_ROOT, file_name: str = None) -> Path:
+    if file_name is None:
+        flow_path, flow_file = resolve_flow_path(get_flow_folder(folder_name, root), check_flow_exist=False)
+        yaml_file = flow_path / flow_file
+    else:
+        yaml_file = get_flow_folder(folder_name, root) / file_name
+
     return yaml_file