+++ src/promptflow-devkit/promptflow/_cli/_pf/entry.py
@@ -2,8 +2,12 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 # pylint: disable=wrong-import-position
-import json
+#
+# PYTHON_ARGCOMPLETE_OK
+
 import time
+
+import argcomplete
 
 from promptflow._cli._pf._experiment import add_experiment_parser, dispatch_experiment_commands
 from promptflow._cli._utils import _get_cli_activity_name, cli_exception_and_telemetry_handler
@@ -28,13 +32,7 @@ from promptflow._cli._pf._trace import add_trace_parser, dispatch_trace_cmds  # 
 from promptflow._cli._pf._upgrade import add_upgrade_parser, upgrade_version  # noqa: E402
 from promptflow._cli._pf.help import show_privacy_statement, show_welcome_message  # noqa: E402
 from promptflow._cli._user_agent import USER_AGENT  # noqa: E402
-from promptflow._sdk._utils import (  # noqa: E402
-    get_promptflow_core_version,
-    get_promptflow_devkit_version,
-    get_promptflow_sdk_version,
-    get_promptflow_tracing_version,
-    print_pf_version,
-)
+from promptflow._sdk._utils import print_pf_version, print_promptflow_version_dict_string  # noqa: E402
 from promptflow._utils.logger_utils import get_cli_sdk_logger  # noqa: E402
 from promptflow._utils.user_agent_utils import setup_user_agent_to_operation_context  # noqa: E402
 
@@ -119,6 +117,8 @@ def get_parser_args(argv):
     add_service_parser(subparsers)
     add_upgrade_parser(subparsers)
 
+    argcomplete.autocomplete(parser)
+
     return parser.prog, parser.parse_args(argv)
 
 
@@ -138,24 +138,7 @@ def main():
     """Entrance of pf CLI."""
     command_args = sys.argv[1:]
     if len(command_args) == 1 and command_args[0] == "version":
-        version_dict = {"promptflow": get_promptflow_sdk_version()}
-        # check tracing version
-        version_tracing = get_promptflow_tracing_version()
-        if version_tracing:
-            version_dict["promptflow-tracing"] = version_tracing
-        # check core version
-        version_core = get_promptflow_core_version()
-        if version_core:
-            version_dict["promptflow-core"] = version_core
-        # check devkit version
-        version_devkit = get_promptflow_devkit_version()
-        if version_devkit:
-            version_dict["promptflow-devkit"] = version_devkit
-
-        version_dict_string = (
-            json.dumps(version_dict, ensure_ascii=False, indent=2, sort_keys=True, separators=(",", ": ")) + "\n"
-        )
-        print(version_dict_string)
+        print_promptflow_version_dict_string()
         return
     if len(command_args) == 0:
         # print privacy statement & welcome message like azure-cli