+++ src/promptflow/tests/sdk_cli_test/e2etests/test_cli.py
@@ -19,7 +19,7 @@ import mock
 import pytest
 
 from promptflow._cli._pf.entry import main
-from promptflow._constants import LINE_NUMBER_KEY, PF_USER_AGENT
+from promptflow._constants import FLOW_FLEX_YAML, LINE_NUMBER_KEY, PF_USER_AGENT
 from promptflow._sdk._constants import LOGGER_NAME, SCRUBBED_VALUE, ExperimentStatus
 from promptflow._sdk._errors import RunNotFoundError
 from promptflow._sdk.operations._local_storage_operations import LocalStorageOperations
@@ -69,7 +69,7 @@ class TestCli:
     def test_pf_version(self, capfd):
         run_pf_command("--version")
         out, _ = capfd.readouterr()
-        assert "0.0.1\n" in out
+        assert "0.0.1" in out
 
     def test_basic_flow_run(self, capfd) -> None:
         # fetch std out
@@ -298,6 +298,42 @@ class TestCli:
         with open(log_path, "r") as f:
             log_content = f.read()
         assert previous_log_content not in log_content
+
+        # Test flow test with simple input file
+        run_pf_command(
+            "flow",
+            "test",
+            "--flow",
+            f"{FLOWS_DIR}/web_classification",
+            "--inputs",
+            f"{DATAS_DIR}/webClassification1.jsonl",
+        )
+        output_path = Path(FLOWS_DIR) / "web_classification" / ".promptflow" / "flow.output.json"
+        assert output_path.exists()
+
+        # Test flow test with invalid simple input file
+        with pytest.raises(ValueError) as ex:
+            run_pf_command(
+                "flow",
+                "test",
+                "--flow",
+                f"{FLOWS_DIR}/web_classification",
+                "--inputs",
+                f"{DATAS_DIR}/invalid_path.json",
+            )
+        assert "Cannot find inputs file" in ex.value.args[0]
+
+        # Test flow test with invalid file extension
+        with pytest.raises(ValueError) as ex:
+            run_pf_command(
+                "flow",
+                "test",
+                "--flow",
+                f"{FLOWS_DIR}/web_classification",
+                "--inputs",
+                f"{DATAS_DIR}/logo.jpg",
+            )
+        assert "Only support jsonl or json file as input" in ex.value.args[0]
 
     def test_flow_with_aad_connection(self):
         run_pf_command("flow", "test", "--flow", f"{FLOWS_DIR}/flow_with_aad_connection")
@@ -1596,6 +1632,18 @@ class TestCli:
         image_path = Path(FLOWS_DIR) / "python_tool_with_composite_image" / ".promptflow" / "intermediate"
         assert image_path.exists()
 
+    def test_flow_test_with_openai_vision_image_input_and_output(self):
+        run_pf_command(
+            "flow",
+            "test",
+            "--flow",
+            f"{FLOWS_DIR}/python_tool_with_openai_vision_image",
+        )
+        output_path = Path(FLOWS_DIR) / "python_tool_with_openai_vision_image" / ".promptflow" / "output"
+        assert output_path.exists()
+        image_path = Path(FLOWS_DIR) / "python_tool_with_openai_vision_image" / ".promptflow" / "intermediate"
+        assert image_path.exists()
+
     def test_run_file_with_set(self, pf) -> None:
         name = str(uuid.uuid4())
         run_pf_command(
@@ -2266,6 +2314,46 @@ class TestCli:
         # check run results
         run = pf.runs.get(run_id)
         assert_batch_run_result(run, pf, assert_func)
+
+    def test_pf_flow_save(self, pf):
+        with tempfile.TemporaryDirectory() as temp_dir:
+            run_pf_command(
+                "flow",
+                "save",
+                "--path",
+                temp_dir,
+                "--entry",
+                "hello:hello_world",
+                "--code",
+                f"{EAGER_FLOWS_DIR}/../functions/hello_world",
+            )
+            assert os.listdir(temp_dir) == [FLOW_FLEX_YAML, "hello.py"]
+            content = load_yaml(Path(temp_dir) / FLOW_FLEX_YAML)
+            assert content == {
+                "entry": "hello:hello_world",
+                "inputs": {
+                    "text": {
+                        "type": "string",
+                    }
+                },
+                "outputs": {
+                    "output": {
+                        "type": "string",
+                    }
+                },
+            }
+            os.unlink(Path(temp_dir) / FLOW_FLEX_YAML)
+            run_pf_command(
+                "flow",
+                "save",
+                "--entry",
+                "hello:hello_world",
+                cwd=temp_dir,
+            )
+            # __pycache__ will be created when inspecting the module
+            assert os.listdir(temp_dir) == [FLOW_FLEX_YAML, "hello.py", "__pycache__"]
+            new_content = load_yaml(Path(temp_dir) / FLOW_FLEX_YAML)
+            assert new_content == content
 
 
 def assert_batch_run_result(run, pf, assert_func):